+++ packages/botonic-plugin-flow-builder/src/content-fields/flow-handoff.tsx
@@ -1,10 +1,5 @@
 import { HandOffBuilder, HelpdeskEvent, isDev, isWebchat } from '@botonic/core'
-import {
-  ActionRequest,
-  Multichannel,
-  Text,
-  WebchatSettings,
-} from '@botonic/react'
+import { ActionRequest, WebchatSettings } from '@botonic/react'
 import React from 'react'
 
 import { FlowBuilderApi } from '../api'
@@ -71,24 +66,12 @@ export class FlowHandoff extends ContentFieldsBase {
       handOffBuilder.withExtraData({
         language,
       })
-      this.isTestIntegration = request.session.is_test_integration
+
       await handOffBuilder.handOff()
     }
   }
 
   toBotonic(id: string, request: ActionRequest): JSX.Element {
-    if (this.isTestIntegration) {
-      return (
-        <Multichannel key={this.id}>
-          <Text>
-            _**HANDOFF IN PREVIEW**_ {'\n'}ℹ️ _At this point, a new case would
-            be created in {this.queue?.name} queue. To continue with the
-            preview, a case resolved scenario will be simulated._
-          </Text>
-        </Multichannel>
-      )
-    }
-
     return isDev(request.session) || isWebchat(request.session) ? (
       <WebchatSettings key={id} enableUserInput={true} />
     ) : (