+++ packages/botonic-plugin-flow-builder/src/content-fields/hubtype-fields/intent.ts
@@ -1,6 +1,7 @@
 import { HtBaseNode, HtInputLocale, HtTextLocale } from './common'
 import { HtNodeWithContentType } from './node-types'
 
+// TODO: Remove this because frontend no allow create intents babel
 export interface HtIntentNode extends HtBaseNode {
   type: HtNodeWithContentType.INTENT
   content: {