+++ packages/botonic-plugin-flow-builder/tests/helpers/utils.ts
@@ -1,6 +1,7 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 import {
   Input,
+  PluginPostRequest,
   PluginPreRequest,
   PROVIDER,
   ProviderType,
@@ -13,13 +14,18 @@ import BotonicPluginFlowBuilder, {
   FlowBuilderActionProps,
   FlowContent,
 } from '../../src'
-import { KnowledgeBaseFunction, TrackEventFunction } from '../../src/types'
+import {
+  InShadowingConfig,
+  KnowledgeBaseFunction,
+  TrackEventFunction,
+} from '../../src/types'
 
 interface FlowBuilderOptions {
   flow: any
   locale?: string
   trackEvent?: TrackEventFunction
   getKnowledgeBaseResponse?: KnowledgeBaseFunction
+  inShadowing?: Partial<InShadowingConfig>
 }
 
 export function createFlowBuilderPlugin({
@@ -27,6 +33,7 @@ export function createFlowBuilderPlugin({
   locale = 'en',
   trackEvent,
   getKnowledgeBaseResponse,
+  inShadowing,
 }: FlowBuilderOptions) {
   const flowBuilderPlugin = new BotonicPluginFlowBuilder({
     flow,
@@ -34,6 +41,7 @@ export function createFlowBuilderPlugin({
     getAccessToken: () => 'fake_token',
     trackEvent,
     getKnowledgeBaseResponse,
+    inShadowing,
   })
 
   // @ts-ignore
@@ -52,6 +60,7 @@ interface RequestArgs {
   provider?: ProviderType
   isFirstInteraction?: boolean
   extraData?: any
+  shadowing?: boolean
 }
 
 export function createRequest({
@@ -60,6 +69,7 @@ export function createRequest({
   provider = PROVIDER.WEBCHAT,
   isFirstInteraction = false,
   extraData = {},
+  shadowing = false,
 }: RequestArgs): PluginPreRequest {
   return {
     session: {
@@ -68,6 +78,7 @@ export function createRequest({
       organization_id: 'orgIdTest',
       bot: { id: 'bid1' },
       user: { provider, id: 'uid1', extra_data: extraData },
+      _shadowing: shadowing,
       __retries: 0,
       _access_token: 'fake_access_token',
       _hubtype_api: 'https://api.hubtype.com',
@@ -81,6 +92,13 @@ export function createRequest({
     },
     lastRoutePath: '',
     plugins,
+    getString: (_stringId: string) => '',
+    setLocale: (_locale: string) => {
+      return
+    },
+    params: {},
+    defaultDelay: 0,
+    defaultTyping: 0,
   }
 }
 
@@ -113,6 +131,7 @@ export async function createFlowBuilderPluginAndGetContents({
 }: FlowBuilderPluginAndGetContentsArgs): Promise<{
   contents: FlowContent[]
   request: PluginPreRequest
+  flowBuilderPluginPost: (request: PluginPostRequest) => void
 }> {
   const flowBuilderPlugin = createFlowBuilderPlugin(flowBuilderOptions)
 
@@ -129,5 +148,9 @@ export async function createFlowBuilderPluginAndGetContents({
     flowBuilderPlugin
   )
 
-  return { contents, request }
+  const flowBuilderPluginPost = (request: PluginPostRequest) => {
+    flowBuilderPlugin.post(request)
+  }
+
+  return { contents, request, flowBuilderPluginPost }
 }